// src/types.ts
export type UserRole = 'management' | 'teacher';
export type MediumType = 'English' | 'Semi-English' | 'Urdu';
export type ExamType = 'Unit Test' | 'Midterm' | 'Final' | 'PreBoard';
export type ExamStatus = 'Draft' | 'Active' | 'Completed';
export type SyncStatus = 'idle' | 'syncing' | 'synced' | 'error' | 'offline';

export interface AuthUser {
  id: string;
  name: string;
  role: UserRole;
}